KS-2-5 - привел оформленеие кода в порядок - убрал перевод в дробные части при расчете amount

+++ src/main/kotlin/lesson_2/lesson2_task5.kt
@@ -7,7 +7,8 @@ fun main() {
     val depositAmount = 70000
     val interestRate = 16.7
     val depositTermYears = 20
-    val amount = depositAmount.toDouble() * (1 + interestRate / 100).pow(depositTermYears.toDouble())
+
+    val amount = depositAmount * (1 + interestRate / 100).pow(depositTermYears)
 
     println("%.3f".format(amount))
 }
